+++ functional/test/test_util.py
@@ -1,7 +1,11 @@
 from __future__ import absolute_import
 
 import unittest
-from functional.util import ReusableFile
+from collections import namedtuple
+
+from functional.util import ReusableFile, is_namedtuple
+
+Data = namedtuple('Tuple', 'x y')
 
 
 class TestUtil(unittest.TestCase):
@@ -12,3 +16,9 @@ class TestUtil(unittest.TestCase):
         iter_1 = iter(license_file_lf)
         iter_2 = iter(license_file_lf)
         self.assertEqual(list(iter_1), list(iter_2))
+
+    def test_is_namedtuple(self):
+        self.assertTrue(is_namedtuple(Data(1, 2)))
+        self.assertFalse(is_namedtuple((1, 2, 3)))
+        self.assertFalse(is_namedtuple([1, 2, 3]))
+        self.assertFalse(is_namedtuple(1))